+++ via.go
@@ -19,7 +19,6 @@ const (
 	viaReboot = "Reboot"
 	viaReset  = "Reset"
 )
-
 // VIA Struct that defines general parameters needed for any VIA
 type Via struct {
 	Address  string
@@ -27,7 +26,7 @@ type Via struct {
 	Password string
 	Logger   Logger
 }
-
+/*
 // command: Struct used to build the XML commands that need to be sent to the VIA
 type command struct {
 	XMLName  xml.Name `xml:"P"`
@@ -45,7 +44,9 @@ type command struct {
 	Param9   string   `xml:"P9,omitempty"`
 	Param10  string   `xml:"P10,omitempty"`
 }
+*/
 
+// These functions fulfill the DSP driver requirements 
 // GetVolumeByBlock: opening a connection with the VIAs and then return the volume for the device
 func (v *Via) GetVolumeByBlock(ctx context.Context, block string) (int, error) {
 	return v.GetVolume(ctx)
@@ -66,142 +67,29 @@ func (v *Via) SetMutedByBlock(ctx context.Context, block string, muted bool) err
 	return errors.New("Error setting mute status of VIA, Feature not supported")
 }
 
-func (v *Via) GetVolume(ctx context.Context) (int, error) {
-	var cmd command
-	cmd.Command = "Vol"
-	cmd.Param1 = "Get"
-
-	log.L.Infof("Sending command to get VIA Volume to %s", v.Address)
-	// Note: Volume Get command in VIA API doesn't have any error handling so it only returns Vol|Get|XX or nothing
-	// I am still checking for errors just in case something else fails during execution
-	vollevel, err := v.SendCommand(ctx, command)
-	if err != nil {
-
-	}
-
-	return v.volumeParse(vollevel)
-}
-
 // GetInfo: needed by the DSP drivers implementation.  Will get hardware information
 func (v *Via) GetInfo(ctx context.Context) (interface{}, error) {
 	var info interface{}
 	return info, fmt.Errorf("GetInfo has not been implemented in this version of the driver")
 }
+// End of DSP Driver requirements
 
-// volumeParse parser to pull out the volume level from the VIA API returned string
-func (v *Via) volumeParse(vollevel string) (int, error) {
-	re := regexp.MustCompile("[0-9]+")
-	vol := re.FindString(vollevel)
-
-	vfin, err := strconv.Atoi(vol)
+func getConnection(address string) (*net.TCPConn, error) {
+	radder, err := net.ResolveTCPAddr("tcp", address+":9982")
 	if err != nil {
-		err = fmt.Errorf("Error converting response: %s", err.Error())
-		log.L.Infof("%s", err.Error())
-		return 0, err
+		err = fmt.Errorf("error resolving address : %s", err.Error())
+		log.L.Infof(err.Error())
+		return nil, err
 	}
 
-	return vfin, nil
-}
-
-// Reboot: Reboot a VIA using the API
-func (v *Via) Reboot(ctx context.Context) error {
-	var command command
-	command.Command = viaReboot
-
-	log.L.Infof("Sending command %s to %s", viaReboot, v.Address)
-
-	_, err := v.SendCommand(ctx, command)
+	conn, err := net.DialTCP("tcp", nil, radder)
 	if err != nil {
-		return err
+		err = fmt.Errorf("error dialing address : %s", err.Error())
+		log.L.Infof(err.Error())
+		return nil, err
 	}
 
-	return nil
-}
-
-// Reset: Reset a VIA sessions - Causes VIAAdmin to log out and log back in which can help with some lock up issues.
-func (v *Via) Reset(ctx context.Context) error {
-	var command command
-	command.Command = viaReset
-
-	log.L.Infof("Sending command %s to %s", viaReset, v.Address)
-
-	resp, err := v.SendCommand(ctx, command)
-	if err != nil {
-		return err
-	}
-
-	if strings.Contains(resp, viaReset) && strings.Contains(resp, "1") {
-		return nil
-	}
-
-	return errors.New(fmt.Sprintf("Incorrect response for command. (Response: %s)", resp))
-}
-
-// Get the Room Code and return the current room code as a string
-func (v *VIA) RoomCode(ctx context.Context) (string, error) {
-	var command Command
-	command.Command = "RCode"
-	command.Param1 = "Get"
-	command.Param2 = "Code"
-
-	log.L.Infof("Sending command to get current room code to %s", v.Address)
-	// Note: RCode Get command in VIA API doesn't have any error handling so it only returns RCode|Get|Code|XXXX or nothing
-	resp, err := v.sendCommand(ctx, command)
-	if err != nil {
-		return "", err
-	}
-	split := strings.Split(resp, "|")
-	if len(split) != 4 {
-		return "", fmt.Errorf("Unknown response %v", resp)
-	}
-
-	roomcode := strings.TrimSpace(split[3])
-
-	return roomcode, nil
-}
-
-// SetAlert - Send an alert to the VIA
-func (v *VIA) SetAlert(ctx context.Context, message string) error {
-	var command Command
-	command.Command = "IAlert"
-	command.Param1 = AlertMessage
-	command.Param2 = "0"
-	command.Param3 = "5"
-
-	log.L.Infof("Sending Alert to %v", v.Address)
-
-	log.L.Debugf("Sending an alert message -%s- to %s", AlertMessage, v.Address)
-
-	resp, err := v.SendCommand(ctx, command)
-	if err != nil {
-		return fmt.Errorf("Error in setting alert on %s", v.Address)
-	}
-	sp := strings.Split(resp, "|")
-	s := sp[1]
-	sint := stings.Atoi(s)
-	if sint != 1 {
-		return fmt.Errorf("Alert was not successfully sent, checking settings and try again")
-	}
-
-	return nil
-}
-
-// SetVolume - Used to set the volume on a VIA (Used by both VIA-Control and DSP Driver sets)
-func (v *Via) SetVolume(ctx context.Context, volume int) (string, error) {
-	var command command
-	command.Command = "Vol"
-	command.Param1 = "Set"
-	command.Param2 = strconv.Itoa(volume)
-
-	log.L.Infof("Sending volume set command to %s", v.Address)
-
-	resp, err := v.SendCommand(ctx, command)
-	if err != nil {
-		return "", errors.New(fmt.Sprintf("Error in setting volume on %s", v.Address))
-	}
-
-	return resp, nil
-
+	return conn, nil
 }
 
 // SendCommand opens a connection with <addr> and sends the <command> to the via, returning the response from the via, or an error if one occured.
@@ -314,27 +202,6 @@ func (v *Via) login(ctx context.Context, conn *net.TCPConn) error {
 	return nil
 }
 
-// TODO remove
-/*func (c *Command) writeCommand(conn *net.TCPConn) error {
-	b, err := xml.Marshal(c)
-	if err != nil {
-		return err
-	}
-
-	return conn.Write(b)
-}
-*/
-// AddAuth adds auth onto the command
-// TODO remove
-// changed: Made function Public
-/*
-func (c *Command) addAuth(viaUser string, viaPass string, password bool) {
-	c.Username = viaUser
-	if password {
-		c.Password = viaPass
-	}
-}
-*/
 func getConnection(address string) (*net.TCPConn, error) {
 	radder, err := net.ResolveTCPAddr("tcp", address+":9982")
 	if err != nil {
